Refactor data exchange reportable properties
