+++ src/Domain/EnforcementEntities/ICaseFileManager.cs
@@ -1,4 +1,4 @@
-﻿using AirWeb.Domain.EnforcementEntities.Cases;
+﻿using AirWeb.Domain.EnforcementEntities.CaseFiles;
 using AirWeb.Domain.Identity;
 
 namespace AirWeb.Domain.EnforcementEntities;
