+++ src/EfRepository/DbContext/Configuration/AppDbContextConfiguration.cs
@@ -3,7 +3,7 @@ using AirWeb.Domain.ComplianceEntities.Fces;
 using AirWeb.Domain.ComplianceEntities.WorkEntries;
 using AirWeb.Domain.EnforcementEntities.ActionProperties;
 using AirWeb.Domain.EnforcementEntities.Actions;
-using AirWeb.Domain.EnforcementEntities.Cases;
+using AirWeb.Domain.EnforcementEntities.CaseFiles;
 using AirWeb.Domain.Identity;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
