+++ src/WebApp/Pages/Enforcement/Details.cshtml
@@ -91,7 +91,7 @@
         <dd class="col-sm-8 col-lg-9">@Html.DisplayFor(_ => caseFile.ResponsibleStaff, DisplayTemplate.NameOrPlaceholder)</dd>
         <dt class="col-sm-4 col-lg-3">@Html.DisplayNameFor(_ => caseFile.DiscoveryDate)</dt>
         <dd class="col-sm-8 col-lg-9">@Html.DisplayFor(_ => caseFile.DiscoveryDate, DisplayTemplate.ShortDateOnlyNullable)</dd>
-        @if (caseFile.HasFormalEnforcement)
+        @if (caseFile.HasReportableEnforcement)
         {
             <dt class="col-sm-4 col-lg-3">@Html.DisplayNameFor(_ => caseFile.DayZero)</dt>
             <dd class="col-sm-8 col-lg-9">@Html.DisplayFor(_ => caseFile.DayZero, DisplayTemplate.ShortDateOnlyNullable)</dd>
@@ -128,7 +128,7 @@
     }
 </section>
 
-@if (caseFile.HasFormalEnforcement)
+@if (caseFile.HasReportableEnforcement)
 {
     <section class="mt-3">
         <h3 class="h4">Associated Pollutants & Programs</h3>
